Role + Settings helper classes to remove logic from controllers
