+++ app/Http/Controllers/RoleController.php
@@ -4,43 +4,11 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\RoleRequest;
 use App\Role;
-use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 
 class RoleController extends Controller
 {
-
-    private static ?RoleController $_instance = null;
-
-    private ?Collection $_roles = null;
-    private ?array $_staff_roles = null;
-
-    public static function getInstance(): RoleController
-    {
-        if (self::$_instance == null) {
-            self::$_instance = new RoleController;
-        }
-        return self::$_instance;
-    }
-
-    public function getRoles(string $order = 'DESC'): object
-    {
-        if ($this->_roles == null) {
-            $this->_roles = Role::where('deleted', false)->orderBy('order', $order)->get();
-        }
-
-        return $this->_roles;
-    }
-
-    public function getStaffRoles(): array
-    {
-        if ($this->_staff_roles == null) {
-            $this->_staff_roles = Role::select('id', 'name')->orderBy('order', 'ASC')->where([['staff', true], ['deleted', false]])->get()->toArray();
-        }
-
-        return $this->_staff_roles;
-    }
 
     public function new(RoleRequest $request)
     {
