+++ app/Http/Controllers/SettingsController.php
@@ -6,62 +6,9 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\DB;
 use App\Settings;
-use Illuminate\Database\Eloquent\Collection;
 
 class SettingsController extends Controller
 {
-
-    private static ?SettingsController $_instance = null;
-
-    private ?float $_gst = null;
-    private ?float $_pst = null;
-    private ?int $_stats_time = null;
-    private ?Collection $_categories = null;
-
-    public static function getInstance(): SettingsController
-    {
-        if (self::$_instance == null) {
-            self::$_instance = new SettingsController;
-        }
-
-        return self::$_instance;
-    }
-
-    public function getGst(): float
-    {
-        if ($this->_gst == null) {
-            $this->_gst = Settings::where('setting', 'gst')->pluck('value')->first();
-        }
-
-        return $this->_gst;
-    }
-
-    public function getPst(): float
-    {
-        if ($this->_pst == null) {
-            $this->_pst = Settings::where('setting', 'pst')->pluck('value')->first();
-        }
-
-        return $this->_pst;
-    }
-
-    public function getStatsTime(): float
-    {
-        if ($this->_stats_time == null) {
-            $this->_stats_time = Settings::where('setting', 'stats_time')->pluck('value')->first();
-        }
-
-        return $this->_stats_time;
-    }
-
-    public function getCategories(): Collection
-    {
-        if ($this->_categories == null) {
-            $this->_categories = Settings::where('setting', 'category')->orderBy('value')->get();
-        }
-
-        return $this->_categories;
-    }
 
     public static function editStatsTime(Request $request)
     {
