+++ app/Http/Controllers/StatisticsPageController.php
@@ -2,12 +2,14 @@
 
 namespace App\Http\Controllers;
 
+use App\Helpers\SettingsHelper;
+
 class StatisticsPageController extends Controller
 {
 
     public function view()
     {
-        return view('pages.statistics.statistics', ['stats_time' => SettingsController::getInstance()->getStatsTime()]);
+        return view('pages.statistics.statistics', ['stats_time' => SettingsHelper::getInstance()->getStatsTime()]);
     }
     
 }
