+++ resources/views/pages/products/form.blade.php
@@ -1,6 +1,6 @@
 @php
 
-use App\Http\Controllers\SettingsController;
+use App\Helpers\SettingsHelper;
 use App\Product;
 $product = Product::find(request()->route('id'));
 @endphp
@@ -63,7 +63,7 @@ $product = Product::find(request()->route('id'));
                     <div class="select">
                         <select name="category" required>
                             {{!! !isset($product->category) ? "<option value=\"\" disabled selected>Select Category...</option>" : '' !!}}
-                            @foreach(SettingsController::getInstance()->getCategories() as $category)
+                            @foreach(SettingsHelper::getInstance()->getCategories() as $category)
                                 <option value="{{ $category->value }}"
                                     {{ (!is_null($product) && $product->category == $category->value) || old('category') == $category->value  ? 'selected' : '' }}>
                                     {{ ucfirst($category->value) }}
